requirements.txt
pytest~=6.1.2
pytest-html
allure-pytest
PyYAML~=5.3.1
selenium~=3.141.0
zmail~=0.2.8

py~=1.9.0